Use selected images from the collection
They have to be of the same origin server to be added into the canvas.

If we add CORS to the api (at least for the domain artsmia.org) we could
serve them straight from there. (@damnabit)

The URLS should be:

//api.artsmia.org/images/10219/300/medium.jpg
//api.artsmia.org/images/3520/300/medium.jpg
//api.artsmia.org/images/593/300/medium.jpg
//api.artsmia.org/images/98954/300/medium.jpg
//api.artsmia.org/images/8480/300/medium.jpg

Me and @tomborger were talking about having a live search the same as
the one he's working on for the CMS that could pull any image out of the
collection to go into the tabloid. That would be kind of cool.